Move `PieChart` to component file
