+++ packages/synapse-interface/components/StateManagedBridge/BridgeExchangeRateInfo.tsx
@@ -10,6 +10,7 @@ import { formatBigIntToString } from '@/utils/bigint/format'
 import { Chain } from '@/utils/types'
 import { CHAINS_BY_ID } from '@constants/chains'
 import * as CHAINS from '@constants/chains/master'
+import { PieChart } from '../PieChart'
 
 const BridgeExchangeRateInfo = ({ showGasDrop }: { showGasDrop: boolean }) => {
   const [gasDropChainId, setGasDropChainId] = useState<number>(null)
@@ -208,25 +209,3 @@ const getAirdropInDollars = (
 }
 
 export default BridgeExchangeRateInfo
-
-export const PieChart = ({
-  activeAmount,
-  totalAmount,
-}: {
-  activeAmount: number
-  totalAmount: number
-}) => {
-  const calculatedActiveDegrees: number = (activeAmount / totalAmount) * 360
-  const calculatedInactiveDegrees: number = 360 - calculatedActiveDegrees
-
-  return (
-    <div
-      data-test-id="pie-chart"
-      className="w-4 h-4 rounded-[50%]"
-      style={{
-        backgroundImage: `
-          conic-gradient(#99E6FF 0deg, #99E6FF ${calculatedActiveDegrees}deg, #343036 ${calculatedActiveDegrees}deg, #343036 ${calculatedInactiveDegrees}deg)`,
-      }}
-    ></div>
-  )
-}
